project02
Project 02 for TeamTreeHouse Front-End TechDegree.

Issues I need help with:

1. Mobile view: mysterious white space at the very top and bottom of
the page.

2. Tablet and Desktop views: The menu at the bottom, I changed it by
using the CSS rule of “visibility”, but this is not the best solution
because you can see that there is a space to the right of the menu.
That’s the hidden “back to top” element.

VALIDATIONS:
html: getting errors saying I can’t use a <div> inside a <ul>

css: the entire desktop code has an error (don’t know which)

This is not my final version, I just want to fix these bugs :)
Thanks!
Paula